route protection & custom video player

+++ controllers/videoController.js
@@ -37,6 +37,7 @@ export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "Upload" });
 
 export const postUpload = async (req, res) => {
+  console.log(req.user._id);
   const {
     body: { title, description },
     file: { path },
@@ -45,7 +46,10 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user._id,
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -54,7 +58,7 @@ export const videoDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
     // video == video:video (진화된 js 작성법)
   } catch (error) {
@@ -68,7 +72,12 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit [${video.title}]`, video });
+    if (video.creator.toString() !== req.user._id.toString()) {
+      throw Error();
+      // try에서 error를 발생시키면 자동적으로 catch로 이동시킴
+    } else {
+      res.render("editVideo", { pageTitle: `Edit [${video.title}]`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -92,7 +101,13 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator.toString() !== req.user._id.toString()) {
+      throw Error();
+      // try에서 error를 발생시키면 자동적으로 catch로 이동시킴
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
